added class Direction type enum
